==> Notebook-Class-exercises/Step-2-Describe-Data.ipynb <==
# Step 2 - Describe Data 

This notebook provides the Python code for opening Covid Total Cases file and computing descriptive statistics.  

Students will be developing a similar notebook for total deaths.  The corresponding notebook is included in the answer section.

#  		Import Libraries

import os
import pandas as pd
import numpy as np

# Set Input Folder

Depending on the Operating System you are using the file access may differ. 
Choose your operating system by setting its value to True and keep the rest False

using_Google_colab = False
using_Anaconda_on_Mac_or_Linux_or_Azure = True
using_Anaconda_on_windows = False

if using_Google_colab:
    from google.colab import drive
    drive.mount('/content/drive')

if using_Google_colab:
    dir_input = "/content/drive/MyDrive/COVID_Project/input"
if using_Anaconda_on_Mac_or_Linux_or_Azure:
    dir_input = "../input"
if using_Anaconda_on_windows:
    dir_input = r"..\input"   

#  		Q 2-2 Open USA Facts Data Source and list data sets

path_USA_Facts = os.path.join(dir_input, "USA_Facts")
path_USA_Facts

files_USA_Facts = os.listdir(os.path.join(dir_input, "USA_Facts"))
files_USA_Facts

# Q 2-2 Understand number of files included in this data source

len(files_USA_Facts)

#  Q 2-3 Open “covid_confirmed_usafacts” dataset file from USA Facts Data Source

path_USA_Facts = os.path.join(dir_input, "USA_Facts", "covid_confirmed_usafacts")
path_USA_Facts

df_covid_confirmed = pd.read_csv(os.path.join(dir_input, "USA_Facts", "covid_confirmed_usafacts.csv"))

# Q 2-3	Review fields included in this  Data Set - “covid_confirmed_usafacts” 

df_covid_confirmed.columns

# Q 2-3	Find number of rows and columns in this  Data Set - “covid_confirmed_usafacts” 

df_covid_confirmed.shape


# Q 2-3 Sample data contents included in “covid_confirmed_usafacts” dataset

df_covid_confirmed

# Q 2-4 Determine basic statistics on “covid_confirmed_usafacts” dataset file from USA Facts Data Source

df_covid_confirmed.describe()

df_covid_confirmed.describe()[['2020-12-18', '2020-12-19', '2020-12-20']]
