==> Notebook-Class-exercises/Step-3-Prepare-Data-Task-0.ipynb <==
import pandas as pd
from datetime import date

population =50
type(population)

average_cases = 31.7
type(average_cases)

countyName = "Orange County"
type(countyName)

days_of_week = ["Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"]
type(days_of_week)

column_list = {"column 1": "Date", "Column 2": "Incremental Cases"}
type(column_list)

mask_policy = False
type(mask_policy)

from datetime import date
today = date(2020, 12, 29).isoformat()
print(today)
