==> Notebook-Class-exercises/Step-3-Prepare-Data-Task-4-County.ipynb <==
## Import Library

import os
import pandas as pd
from datetime import date


## Set up Environment Flag

using_Google_colab = False
using_Anaconda_on_Mac_or_Linux = True
using_Anaconda_on_windows = False

if using_Google_colab:
    dir_input = "/content/drive/MyDrive/COVID_Project/input"
    dir_output = "/content/drive/MyDrive/COVID_Project/output"
if using_Anaconda_on_Mac_or_Linux:
    dir_input = "../input"
    dir_output = "../output"
if using_Anaconda_on_windows:
    dir_input = r"..\input"
    dir_output = r"..\output" 

## Connect to Google Drive

if using_Google_colab:
    from google.colab import drive
    drive.mount('/content/drive')

Data Prep for USA_Facts confirmed cases at county level

Read file

df_confirmed_cases = pd.read_csv(os.path.join(dir_input, 
                                              'USA_Facts', 
                                              'covid_confirmed_usafacts.csv'))
df_confirmed_cases = df_confirmed_cases.astype({'countyFIPS': str}).astype({'StateFIPS': str})
df_confirmed_cases

Now I need to transpose so all dates are in rows


df_confirmed_cases_by_county_by_date = df_confirmed_cases.melt(id_vars=['County Name',
                                                                        'countyFIPS', 
                                                                        'State', 
                                                                        'StateFIPS'], 
                                                               var_name='Date', 
                                                               value_name='Total Cases')
df_confirmed_cases_by_county_by_date

df_confirmed_cases_by_county_by_date = df_confirmed_cases_by_county_by_date.astype({'Date': 'datetime64[ns]'})
df_sorted_confirmed_cases = df_confirmed_cases_by_county_by_date.sort_values(by=['countyFIPS','Date'])
df_sorted_confirmed_cases

Now shift by one to get previous day cases and compute incremental cases

df_sorted_confirmed_cases['Inc_Cases'] = df_sorted_confirmed_cases.groupby('countyFIPS')['Total Cases'].apply(
    lambda x: x - x.shift(1))
df_sorted_confirmed_cases

Now compute 7 days rolling average

df_sorted_confirmed_cases['cases_moving_avg'] = df_sorted_confirmed_cases.groupby('countyFIPS')['Inc_Cases'].apply(
    lambda x: (x + x.shift(1) + x.shift(2) + x.shift(3) + x.shift(4) + x.shift(5) + x.shift(6))/7)
df_sorted_confirmed_cases

def direction_of_spread(x):
    if x['Inc Cases'] > x['moving_avg']: 
        return 'increasing'
    else: 
        return'decreasing'

df_sorted_confirmed_cases['difference'] = df_sorted_confirmed_cases['Inc_Cases'] - df_sorted_confirmed_cases[
    'cases_moving_avg']
df_sorted_confirmed_cases['direction']=df_sorted_confirmed_cases['difference'].apply(
    lambda x: 'increasing' if x>0 else 'not increasing')
df_oc_confirmed_cases = df_sorted_confirmed_cases[df_sorted_confirmed_cases['countyFIPS'] == '6059']

df_sorted_confirmed_cases.to_csv(os.path.join(dir_output, 'confirmed_cases_by_county.csv'))

df_oc_confirmed_cases.to_csv(os.path.join(dir_output, 'confirmed_cases_orange_county_ca.csv'))

Data prep for COVID deaths by County

Read file

df_covid_deaths = pd.read_csv(os.path.join(dir_input, "USA_Facts", "covid_deaths_usafacts.csv"))
df_covid_deaths = df_covid_deaths.astype({'countyFIPS': str}).astype({'StateFIPS': str})
df_covid_deaths

Now group the data by State

Now I need to transpose so all dates are in rows

df_covid_deaths_by_county_by_date = df_covid_deaths.melt(id_vars=['State',
                                                                  'StateFIPS', 
                                                                  'County Name', 
                                                                  'countyFIPS'], 
                                                         var_name='Date', 
                                                         value_name='Total Deaths')
df_covid_deaths_by_county_by_date

df_covid_deaths_by_county_by_date = df_covid_deaths_by_county_by_date.astype({'Date': 'datetime64[ns]'})
df_sorted_covid_deaths = df_covid_deaths_by_county_by_date.sort_values(by=['countyFIPS', 'Date'])
df_sorted_covid_deaths

Now shift by one to get previous day deaths and compute incremental deaths

df_sorted_covid_deaths['Inc Deaths'] = df_sorted_covid_deaths.groupby('countyFIPS')['Total Deaths'].apply(
    lambda x: x - x.shift(1))
df_sorted_covid_deaths

Now compute 7 days rolling average

df_sorted_covid_deaths['death moving_avg'] = df_sorted_covid_deaths.groupby('countyFIPS')['Inc Deaths'].apply(
    lambda x: (x + x.shift(1) + x.shift(2) + x.shift(3) + x.shift(4) + x.shift(5) + x.shift(6))/7)
df_sorted_covid_deaths

df_sorted_covid_deaths.to_csv(os.path.join(dir_output, 'covid_deaths_by_county.csv'))

Create partial analytics_base_table with confirmed cases and deaths


df_partial_abt_by_county = pd.merge(df_sorted_confirmed_cases, df_sorted_covid_deaths, 
                                    on=['countyFIPS', 'Date'], 
                                    suffixes=('', '_DROP'), 
                                    how='inner').filter(regex='^(?!.*_DROP)')
df_partial_abt_by_county

df_partial_abt_by_county[df_partial_abt_by_county['State'] == "CA"]['County Name'].unique()

df_partial_abt_by_county[df_partial_abt_by_county['County Name'] == 'Los Angeles County '].count()

df_partial_abt_by_county[df_partial_abt_by_county['County Name'] == 'Los Angeles County '].Date.min()

df_partial_abt_by_county[df_partial_abt_by_county['County Name'] == 'Los Angeles County '].Date.max()
