==> Notebook-Class-exercises/Step-3-Prepare-Data-Task-7-Merge-Datasets.ipynb <==
# Step 3 - Prepare Data - Task 7 - Merge Dataset

## Import libraries

import os
import pandas as pd
from datetime import date


## Set up environment flag

using_Google_colab = False
using_Anaconda_on_Mac_or_Linux = True
using_Anaconda_on_windows = False

if using_Google_colab:
    dir_input = "/content/drive/MyDrive/COVID_Project/input"
    dir_output = "/content/drive/MyDrive/COVID_Project/output"
if using_Anaconda_on_Mac_or_Linux:
    dir_input = "../input"
    dir_output = "../output"    
if using_Anaconda_on_windows:
    dir_input = r"..\input"  
    dir_output = r"..\output"  

## If using Google colab, get connected to google drive

if using_Google_colab:
    from google.colab import drive
    drive.mount('/content/drive')


## PD 7.1 Activity 1 - Upload county level covid cases data by county

df_sorted_confirmed_cases_county = pd.read_csv(os.path.join(dir_output, 'confirmed_cases_by_county.csv'))
df_sorted_confirmed_cases_county = df_sorted_confirmed_cases_county.astype({'countyFIPS': int, 
                                                                            'StateFIPS': int,
                                                                            'Date': 'datetime64[ns]'})
df_sorted_confirmed_cases_county

## PD 7.1 Activity 2 - Upload county level covid deaths by county

df_sorted_covid_deaths_county = pd.read_csv(os.path.join(dir_output, 'covid_deaths_by_county.csv'))
df_sorted_covid_deaths_county = df_sorted_covid_deaths_county.astype({'countyFIPS': int, 
                                                                            'StateFIPS': int,
                                                                            'Date': 'datetime64[ns]'})
df_sorted_covid_deaths_county

## PD 7.2 Activity 3 - Merge covid cases and deaths data for each county and date

df_partial_abt_by_county = pd.merge(df_sorted_confirmed_cases_county, df_sorted_covid_deaths_county, 
                                    on=['StateFIPS','countyFIPS', 'Date'], 
                                    suffixes=('', '_DROP'), 
                                    how='inner').filter(regex='^(?!.*_DROP)')
df_partial_abt_by_county

## PD 7.3 Activity 4 - Upload county level Google social mobility data

df_google_mobility_data = pd.read_csv(os.path.join(dir_input, 'Google', 'Region_Mobility_Report_CSVs', '2020_US_Region_Mobility_Report.csv'))
df_google_mobility_data = df_google_mobility_data.astype({'date': 'datetime64[ns]'})
df_google_mobility_data

## PD 7.3 Activity 5 - Understand Google Mobility data

df_google_mobility_data.columns


df_partial_abt_by_county[df_partial_abt_by_county['County Name'] == 'Los Angeles County '].count()

df_google_mobility_data[df_google_mobility_data['sub_region_2'] == 'Los Angeles County'].count()

df_google_mobility_data[df_google_mobility_data['sub_region_2'] == 'Los Angeles County'].date.min()

df_google_mobility_data[df_google_mobility_data['sub_region_2'] == 'Los Angeles County'].date.max()

df_partial_abt_by_county[df_partial_abt_by_county['County Name'] == 'Los Angeles County '].Date.min()

df_partial_abt_by_county[df_partial_abt_by_county['County Name'] == 'Los Angeles County '].Date.max()

## PD 7.3 Activity 6: Reformat Google Mobility data


df_google_mobility_data_clean = df_google_mobility_data.dropna(subset=['census_fips_code'])
df_google_mobility_data_clean = df_google_mobility_data_clean.astype({'census_fips_code': int})
df_google_mobility_data_clean


## PD 7.4 Activity 7 - Merge Covid Cases, deaths and Google Data

df_abt_by_county = pd.merge(df_partial_abt_by_county, 
                            df_google_mobility_data, left_on=['countyFIPS', 'Date'], 
                            right_on=['census_fips_code', 'date'], 
                            suffixes=('', '_DROP'), 
                            how='left').filter(regex='^(?!.*_DROP)')
df_abt_by_county

## PD7.4 Activity 8: Save County Data as Analytics Base Table

df_abt_by_county.to_csv(os.path.join(dir_output, 'abt_by_county.csv'))

df_abt_LA_county = df_abt_by_county[df_abt_by_county['County Name'] == 'Los Angeles County ']
df_abt_LA_county

df_abt_LA_county.to_csv(os.path.join(dir_output, 'abt_LA_county.csv'))

df_abt_Orange_county_CA = df_abt_by_county[df_abt_by_county['countyFIPS'] == 6059]
df_abt_Orange_county_CA

df_abt_Orange_county_CA.to_csv(os.path.join(dir_output, 'abt_Orange_county.csv'))
