==> Notebook-Class-exercises/Step-3-1 Prepare-Data-Select.ipynb <==
# Step 3-1 - Prepare Data - Select

This notebook provides the Python code for taking Google Mobility Data at county level, concatenating data from multiple years and then selecting a subset of data with appropriate changes to column names and data types.  Its results are stored in an intermediate file for rest of exercises.

Students will be developing a similar notebook for Canadian Google Mobility data.  The corresponding notebook is included in the answer section.

# Q 3-1 

3-1-1. Read Google Mobility data

3-1-2. select:
- sub_region_1 – rename state name
- sub_region_2 – rename county name
- census_fips_code –rename countyFIPS
- date
- mobility indicators

3-1-3. Change data types
- FIPS code

# Import Libraries

import os
import pandas as pd

# Set Input Folder

Depending on the Operating System you are using the file access may differ. 
Choose your operating system by setting its value to True and keep the rest False

using_Google_colab = False
using_Anaconda_on_Mac_or_Linux_or_Azure = True
using_Anaconda_on_windows = False

if using_Google_colab:
    from google.colab import drive
    drive.mount('/content/drive')

if using_Google_colab:
    dir_input = "/content/drive/MyDrive/COVID_Project/input"
    dir_output = "/content/drive/MyDrive/COVID_Project/output"
if using_Anaconda_on_Mac_or_Linux_or_Azure:
    dir_input = "../input"
    dir_output = "../output"
if using_Anaconda_on_windows:
    dir_input = r"..\input"   
    dir_output = r"..\output" 

### 3-1-1 Read files
- Read google_mobility_data_2020
- Read google_mobility_data_2021
- Read google_mobility_data_2022
- Concat the three files to create google_mobility_data

df_google_mobility_data_2020 = pd.read_csv(os.path.join(dir_input, 
                                                   "Google", 
                                                   "Region_Mobility_Report_CSVs", 
                                                   "2020_US_Region_Mobility_Report.csv"))
df_google_mobility_data_2020

df_google_mobility_data_2021 = pd.read_csv(os.path.join(dir_input, 
                                                   "Google", 
                                                   "Region_Mobility_Report_CSVs", 
                                                   "2021_US_Region_Mobility_Report.csv"))
df_google_mobility_data_2021

df_google_mobility_data_2022 = pd.read_csv(os.path.join(dir_input, 
                                                   "Google", 
                                                   "Region_Mobility_Report_CSVs", 
                                                   "2022_US_Region_Mobility_Report.csv"))
df_google_mobility_data_2022

df_google_mobility_data = pd.concat([df_google_mobility_data_2020, 
                                     df_google_mobility_data_2021, 
                                     df_google_mobility_data_2022], ignore_index=True)
df_google_mobility_data

df_google_mobility_data.columns

### 3-1-2 Select
- sub_region_1
- sub_region_2
- census_fips_code
- date
- Mobility indicators:
-- retail_and_recreation_percent_change_from_baseline
-- grocery_and_pharmacy_percent_change_from_baseline
-- parks_percent_change_from_baseline
-- transit_stations_percent_change_from_baseline
-- workplaces_percent_change_from_baseline
-- residential_percent_change_from_baseline


df_google_mobility_data_selected = df_google_mobility_data[['sub_region_1','sub_region_2','census_fips_code','date',
                                                             'retail_and_recreation_percent_change_from_baseline',
                                                             'grocery_and_pharmacy_percent_change_from_baseline',
                                                             'parks_percent_change_from_baseline',
                                                             'transit_stations_percent_change_from_baseline',
                                                             'workplaces_percent_change_from_baseline',
                                                             'residential_percent_change_from_baseline']]
                                              
df_google_mobility_data_selected

df_google_mobility_data_selected = df_google_mobility_data_selected.rename(columns={"census_fips_code": "countyFIPS"})
df_google_mobility_data_selected

df_google_mobility_data_selected.columns

### 3-1-3 Change data types

Change FIPS code to string after dropping an accidental .0 (see 56045.0 in the last line) added by Python reader
By converting data type to int, it drops the .0 and then it is converted to string

df_google_mobility_data_selected = df_google_mobility_data_selected.dropna(subset=['countyFIPS'])
df_google_mobility_data_selected = df_google_mobility_data_selected.astype({'countyFIPS': int})
df_google_mobility_data_selected = df_google_mobility_data_selected.astype({'countyFIPS': str})
df_google_mobility_data_selected

df_google_mobility_data_selected.to_csv(os.path.join(dir_output,  
                                                   "Google_Mobility_Report.csv"))
