==> Notebook-Class-exercises/Step-3-2 Prepare-Data-Transform.ipynb <==
# Step 3 - Prepare Data - Transform

This notebook provides the Python code for taking Covid total cases at county level and creating state level aggregation.  It demonstrates how you can transform the data.  Its results are stored in an intermediate file for rest of exercises.

Q 3-2 USA Facts data has dates in columns, transform the data so it is indexed by date and has a row for each date for a county.
- State and county as index
- New column - Date
- New column - Total Cases

Students will be developing a similar notebook for total deaths.  The corresponding notebook is included in the answer section.

# Q 3-2 

3-2-1. Read USA Facts data on total cases

3-2-2. Filter LA County so you can check results on a sample

3-2-3. Transform LA County data and check results

3-2-4  Transform data for all counties

3-2-5  Write transformed data in output folder


## Import Libraries

import os
import pandas as pd
from datetime import date


# Set Input Folder

Depending on the Operating System you are using the file access may differ. 
Choose your operating system by setting its value to True and keep the rest False

using_Google_colab = False
using_Anaconda_on_Mac_or_Linux_or_Azure = True
using_Anaconda_on_windows = False

if using_Google_colab:
    dir_input = "/content/drive/MyDrive/COVID_Project/input"
    dir_output = "/content/drive/MyDrive/COVID_Project/output"
if using_Anaconda_on_Mac_or_Linux_or_Azure:
    dir_input = "../input"
    dir_output = "../output"
if using_Anaconda_on_windows:
    dir_input = r"..\input"   
    dir_output = r"..\output" 

## Connect to Google Drive

This step will only be executed if you have set environment flag using_Google_colab to True

if using_Google_colab:
    from google.colab import drive
    drive.mount('/content/drive')

### 3-2-1 Read Total Cases data

Read data
Fix data types for CountyFIPS and StateFIPS

df_total_cases = pd.read_csv(os.path.join(dir_input, "USA_Facts", "covid_confirmed_usafacts.csv"))
df_total_cases = df_total_cases.astype({'countyFIPS': str}).astype({'StateFIPS': str})
df_total_cases

### 3-2-2 Filter data for LA County

df_total_cases_LA = df_total_cases[df_total_cases['County Name'] == 'Los Angeles County ']
df_total_cases_LA

### 3-2-3 Transform LA County data to total cases by date

df_total_cases_LA_by_date = df_total_cases_LA.melt(id_vars=['State', 
                                                            'StateFIPS', 
                                                            'County Name',
                                                            'countyFIPS'],
                                                   var_name='Date', 
                                                   value_name='Total Cases')
df_total_cases_LA_by_date

### 3-2-4 Transform all County data to total cases by date

df_total_county_cases_by_date = df_total_cases.melt(id_vars=['State', 
                                                      'StateFIPS', 
                                                      'County Name',
                                                      'countyFIPS'],
                                             var_name='Date', 
                                             value_name='Total Cases')
df_total_county_cases_by_date

### 3-2-5 Write Transformed data in output folder

df_total_county_cases_by_date.to_csv(os.path.join(dir_output, "total_county_cases_by_date.csv"))
