==> Notebook-Class-exercises/Step-4-Develop-Model-Task-6-Connect-the-Dots -Task7-Graph-Analytics.ipynb <==
### Install Pygeohash library

This library provides functions for computing geohash


## Step 5 - Develop Model - Task 6 - Connect the dots & Task 7 - Graph Analytics - CLASS ASSIGNMENTS

!pip install pygeohash

### Import pygeohash, networkx and Pandas libraries

Pygeohash - functions for converting latitude, longitude to geohash and related distance measurement utilities

Networkx - functions for creating, manipulating and querying open source network graphs 

Pandas - Python functions for table manipuation

import pygeohash as pgh
import networkx as nx
import pandas as pd

### Connect to datasets using Google drive or local files

using_Google_colab = True
using_Anaconda_on_Mac_or_Linux = False
using_Anaconda_on_windows = False

if using_Google_colab:
    from google.colab import drive
    drive.mount('/content/drive')

##DM6.1 Open  Notebook, read Lat, Long and compute Geohash - Activity 1

if using_Google_colab:
  state_location = pd.read_csv('/content/drive/MyDrive/COVID_Project/input/state_lat_long.csv')
if using_Anaconda_on_Mac_or_Linux:
  state_location = pd.read_csv('../input/state_lat_long.csv')
if using_Anaconda_on_windows:
  state_location = pd.read_csv(r'..\input\state_lat_long.csv')
state_location.loc[0:5,]

### Apply a function call to convert Lat, Long to Geohash

def lat_long_to_geohash(lat_long):
  return pgh.encode(lat_long[0], lat_long[1])

state_location['geohash'] = state_location[['latitude', 
                                            'longitude']].apply(lat_long_to_geohash, 
                                                                axis=1)
state_location.iloc[0:10,]

### Truncate geohash to first two characters

state_location['geohash'] = state_location.geohash.str.slice(stop=2)
state_location.iloc[0:10,]

## DM6.2 - Design Graph representaing States and Geohash

### Find neighbors by sorting the states by 2 character geohash codes attached to each state

### Initialize Graph and create state and geohash concepts as nodes

GRAPH_ID = nx.DiGraph()
GRAPH_ID.add_node('state')
GRAPH_ID.add_node('geohash')

### Create a node for each state

state_list = state_location.state.values
for state in state_list:
  GRAPH_ID.add_node(state)
  GRAPH_ID.add_edge('state', state, label='instance')

### Create a list of unique geohash codes and create a node for each geohash

geohash_list = state_location.geohash.values
for geohash in geohash_list:
  GRAPH_ID.add_node(geohash)
  GRAPH_ID.add_edge('geohash', geohash, label='instance')

df_state_geohash = state_location[['state', 'geohash']]
for state_geohash in df_state_geohash.itertuples():
    GRAPH_ID.add_edge(state_geohash.state, state_geohash.geohash, 
                      label='located_at')
    GRAPH_ID.add_edge(state_geohash.geohash, state_geohash.state, 
                      label='locates', 
                      distance=0.0)

## DM6.3 - Which states are in Geohash 9q

### Find geohash associated with California and Naveda


list(GRAPH_ID.neighbors('CA'))

list(GRAPH_ID.neighbors('NV'))

### Find States locsted with geohash '9q'

list(GRAPH_ID.neighbors('9q'))

## DM6.4 Sort the data and find neighbors sharing geohash

### Find states located with geohah for all geohashes

for geohash in GRAPH_ID['geohash']:
  print("Geohash: ", geohash, "States: ", list(GRAPH_ID.neighbors(geohash)))

## Step 4 - Develop Model - Task 7 - Graph Analytics - DM7.1 Activity 1 - Find number of state and geohash nodes in a graph

len(list (GRAPH_ID.neighbors('geohash')))

len(list (GRAPH_ID.neighbors('state')))

## DM7.2 - Find all neighboring states for NY

### Connect neighboring geohash codes if the distance is less than 1,000 km

for geohash_1 in geohash_list:
  for geohash_2 in geohash_list:
    if geohash_1 != geohash_2:
      distance = pgh.geohash_haversine_distance(geohash_1, geohash_2)
      if distance < 1000000:
        GRAPH_ID.add_edge(geohash_1, geohash_2, label='near')

### Find path length from NY to all nodes (states and geohashes)

neighbor_path_length = nx.single_source_dijkstra_path_length(GRAPH_ID, 'NY', weight='distance')
neighbor_path_length

### Make a list of all nodes covered in the path length and then find those nodes which are states and less than or equal to 3 hops

neighbor_states = neighbor_path_length.keys()

state_list = (list (GRAPH_ID.neighbors('state')))
for state in state_list:
  if state in neighbor_states:
    if neighbor_path_length[state] <= 3:
      print(state)


## DM7.3 - Find all neighboring states for each state

for state_1 in state_list:
  neighbor_path_length = nx.single_source_dijkstra_path_length(GRAPH_ID, state_1)
  neighbor_state_list = neighbor_path_length.keys()
  next_door_list = []
  for state_2 in neighbor_state_list:
    if state_1 != state_2:
      if state_2 in state_list:
        if neighbor_path_length[state_2] <=3:
          next_door_list.append(state_2)
  if next_door_list:
    print(state_1, next_door_list)

## DM7.4 - Find path between two states

nx.dijkstra_path(GRAPH_ID, 'NY', 'CA', weight='distance')

nx.dijkstra_path(GRAPH_ID, 'OR', 'CA', weight='distance')

GRAPH_ID.nodes()


nx.single_source_dijkstra_path_length(GRAPH_ID, 'NY')
