==> Notebook-Class-exercises/Step-6-Deploy-Model-Task-1-Scoring-Engine.ipynb <==
## Step 6 – Deploy Model - Task 1. Scoring Engine


###Load Libraries


!pip install sklearn --upgrade

!pip install skl2onnx

!pip install onnxruntime

import pandas as pd
import numpy as np

from sklearn.model_selection import train_test_split 
from sklearn.tree import DecisionTreeClassifier
from skl2onnx import convert_sklearn
from skl2onnx.common.data_types import FloatTensorType
import onnxruntime as rt

### Set up environment and connect to Google Drive

using_Google_colab = False
using_Anaconda_on_Mac_or_Linux = True
using_Anaconda_on_windows = False

if using_Google_colab:
    from google.colab import drive
    drive.mount('/content/drive')

### Upload intermediate classification modeling data for LA County
(see Step-4-Develop-Model-Task-4-Classification notebook for details)

## DM6.1 Open Notebook and upload the LA county merged data - Activity 1


if using_Google_colab:
  merged_LA = pd.read_csv('/content/drive/MyDrive/COVID_Project/output/merged_LA.csv')
if using_Anaconda_on_Mac_or_Linux:
  merged_LA = pd.read_csv('../output/merged_LA.csv')
if using_Anaconda_on_windows:
  merged_LA = pd.read_csv(r'..\output\merged_LA.csv')
merged_LA

## DM6.2 Prepare inputs and outputs for classification model and develop model - Activity 2


### Prepare and apply the model


y_LA = merged_LA['case_direction'].values
X_LA = merged_LA[['delta_retail_recreation', 
               'delta_grocery_pharmacy', 
               'delta_parks', 
               'delta_transit', 
               'delta_workplaces', 
               'delta_residential']].values
X_LA

y_LA

decision_tree = DecisionTreeClassifier()
X_train_LA, X_test_LA, y_train_LA, y_test_LA = train_test_split(X_LA, 
                                                                y_LA, 
                                                                test_size=0.2, 
                                                                random_state=0)
model_LA = decision_tree.fit(X_train_LA, y_train_LA)

## Save Test data into CSV file

df_test_y_LA = pd.DataFrame(y_test_LA, columns=['Y'])
df_test_y_LA

df_test_y_LA.to_csv('../output/test_y_LA.csv')

df_test_X_LA = pd.DataFrame(X_test_LA, columns=['X1', 'X2', 'X3', 'X4', 'X5', 'X6'])
df_test_X_LA

df_test_X_LA.to_csv('../output/test_X_LA.csv')

## DM6.3 - Pack and save model - Activity 3

# Convert into ONNX format

initial_type = [('float_input', FloatTensorType([None, 6]))]
onx = convert_sklearn(decision_tree, initial_types=initial_type)
with open("../output/rf_classification_LA.onnx", "wb") as f:
    f.write(onx.SerializeToString())

### This model can now be tested in Google Colab  - This is your Activity 4 and listed under Task 2 of  Scoring Engine notbook code
