==> Notebook-Class-exercises/Step-6-Deploy-Model-Task-2-Scoring-Engine.ipynb <==
## Step 6- Deploy Model - Task 2 - Deploy Anaconda devevloped model as in Task 1 in Google Colab

!pip install skl2onnx

!pip install onnxruntime

import onnxruntime as rt

import pandas as pd
import numpy as np


from google.colab import drive
drive.mount('/content/drive')

df_test_X_LA = pd.read_csv('/content/drive/MyDrive/COVID_Project/output/test_X_LA.csv', index_col=0)
df_test_y_LA = pd.read_csv('/content/drive/MyDrive/COVID_Project/output/test_y_LA.csv', index_col=0)
df_test_X_LA

df_test_y_LA

y_test_LA = df_test_y_LA.values
X_test_LA = df_test_X_LA.values
X_test_LA

### Load model developed on a different environment (Anaconda as part of Task 1 Notebook) - DM 6.4 - Activity 4

sess = rt.InferenceSession("/content/drive/MyDrive/COVID_Project/output/rf_classification_LA.onnx")
input_name = sess.get_inputs()[0].name
label_name = sess.get_outputs()[0].name
pred_onx_LA = sess.run([label_name], {input_name: X_test_LA.astype(np.float32)})

pred_onx_LA
